Fix LAS/PLY reading functions for Point_set_3
